refactor userMenu style and code
